+++ django_countries/fields.py
@@ -142,11 +142,12 @@ class CountryField(CharField):
     descriptor_class = CountryDescriptor
 
     def __init__(self, *args, **kwargs):
+        self.countries = kwargs.pop('countries', countries)
         self.countries_flag_url = kwargs.pop('countries_flag_url', None)
         self.blank_label = kwargs.pop('blank_label', None)
         kwargs.update({
             'max_length': 2,
-            'choices': countries,
+            'choices': self.countries,
         })
         super(CharField, self).__init__(*args, **kwargs)
 
@@ -184,6 +185,10 @@ class CountryField(CharField):
         """
         name, path, args, kwargs = super(CountryField, self).deconstruct()
         kwargs.pop('choices')
+        if self.countries is not countries:
+            # Include the countries keyword if it's not the default countries
+            # instance.
+            kwargs['countries'] = self.countries
         return name, path, args, kwargs
 
     def get_choices(