+++ django_countries/tests/custom_countries.py
@@ -1,0 +1,6 @@
+from django_countries import Countries
+
+
+countries = Countries(
+    only={'NZ': 'New Zealand', 'NV': 'Neverland'},
+)