+++ django_countries/tests/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django_countries.fields import CountryField
+from django_countries.tests import custom_countries
 
 
 def en_zed():
@@ -12,6 +13,7 @@ class Person(models.Model):
     other_country = CountryField(
         blank=True, countries_flag_url='//flags.example.com/{code}.gif')
     favourite_country = CountryField(default='NZ')
+    fantasy_countries = CountryField(countries=custom_countries.countries)
 
 
 class AllowNull(models.Model):